requirement.txt
altgraph==0.17
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.3
pefile==2019.4.18
Pillow==8.0.1
py-notifier==0.1.1
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqtgraph==0.11.0
pyserial==3.5
python-dateutil==2.8.1
pyusb==1.1.0
pywin32==300
pywin32-ctypes==0.2.0
six==1.15.0
win10toast==0.9
